Create FrameProcessorController and FrameReceiverController and Adapters
