+++ python/src/odin_data/control/frame_receiver_controller.py
@@ -1,0 +1,5 @@
+from odin_data.control.odin_data_controller import OdinDataController
+
+
+class FrameReceiverController(OdinDataController):
+    ...
